some fixes

+++ lib/actions/patient.actions.ts
@@ -1,7 +1,10 @@
 "use server";
 
+import { parseStringify } from "./../utils";
+
 import Patient from "../database/patient.model";
 import { connectToDB } from "../mongoose";
+import Appointment from "../database/appointment.model";
 
 interface CreatePatientParams {
   clerkId: string;
@@ -64,13 +67,37 @@ export const updatePatient = async ({
 export const deletePatient = async (clerkId: string) => {
   try {
     await connectToDB();
+    // Find the patient by clerkId
+    const patient = await Patient.findOne({ clerkId });
+    if (!patient) {
+      throw new Error("Patient not found");
+    }
+
+    // Delete all appointments related to the patient
+    await Appointment.deleteMany({ _id: { $in: patient.appointments } });
+
+    // Delete the patient
     const deletedPatient = await Patient.findOneAndDelete({ clerkId });
     if (!deletedPatient) {
-      throw new Error("Patient not found");
+      throw new Error("Error deleting patient");
     }
     return deletedPatient;
   } catch (error) {
     console.error("Error deleting patient:", error);
     throw new Error("Error deleting patient");
+  }
+};
+
+export const getPatientByClerkId = async (clerkId: string) => {
+  try {
+    await connectToDB();
+    const patient = await Patient.findOne({ clerkId }).populate("appointments");
+    if (!patient) {
+      throw new Error("Patient not found");
+    }
+    return parseStringify(patient);
+  } catch (error) {
+    console.error("Error fetching patient:", error);
+    throw new Error("Error fetching patient");
   }
 };
